Minor Repair
feat(repair): the minor repair is added with an example prefab
Note: the repair script happens a bit differently that the other interactions. Will maybe change how some other interactions work
