+++ Assets/Level/Prefabs/Repair/MinorRepairEXAMPLE.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2555d808b268fc241b7ebb13b8bf35f4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
